requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
absl-py==0.10.0
astunparse==1.6.3
attrs==20.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
bert-for-tf2==0.14.9
cachetools==4.2.1; python_version ~= '3.5'
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cycler==0.10.0
cython==0.29.21; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
dataclasses==0.6
dill==0.3.3; python_version >= '2.6' and python_version != '3.0'
dm-tree==0.1.5
flatbuffers==1.12
future==0.18.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
gast==0.3.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
gin-config==0.4.0
google-api-core[grpc]==1.26.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-api-python-client==1.12.8; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2; python_version >= '3.6'
google-auth==1.27.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-cloud-bigquery==2.9.0; python_version < '3.10' and python_version >= '3.6'
google-cloud-core==1.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-crc32c==1.1.2; python_version >= '3.5'
google-pasta==0.2.0
google-resumable-media==1.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
googleapis-common-protos==1.52.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
grpcio==1.32.0
h5py==2.10.0
httplib2==0.19.0
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
importlib-metadata==3.4.0; python_version < '3.8'
importlib-resources==5.1.0; python_version < '3.9'
joblib==1.0.1; python_version >= '3.6'
kaggle==1.5.10
keras-preprocessing==1.1.2
kiwisolver==1.3.1; python_version >= '3.6'
markdown==3.3.3; python_version >= '3.6'
matplotlib==3.3.4; python_version >= '3.6'
numpy==1.19.5
oauthlib==3.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
opencv-python-headless==4.5.1.48; python_version >= '3.6'
opt-einsum==3.3.0; python_version >= '3.5'
packaging==20.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pandas==1.2.2
params-flow==0.8.2; python_version >= '3.5'
pillow==8.1.0; python_version >= '3.6'
promise==2.3
proto-plus==1.13.0; python_version >= '3.6'
protobuf==3.15.0
psutil==5.8.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
py-cpuinfo==7.0.0
py-params==0.10.2; python_version >= '3.6'
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-slugify==4.0.1
pytz==2021.1
pyyaml==5.4.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
requests-oauthlib==1.3.0
requests==2.25.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rsa==4.7.1; python_version >= '3.6'
scikit-learn==0.24.1
scipy==1.6.1; python_version >= '3.7'
sentencepiece==0.1.95
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tensorboard-plugin-wit==1.8.0
tensorboard==2.4.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tensorflow-addons==0.12.1
tensorflow-datasets==4.2.0; python_version >= '3.6'
tensorflow-estimator==2.4.0
tensorflow-hub==0.11.0
tensorflow-metadata==0.27.0; python_version >= '3.6' and python_version < '4'
tensorflow-model-optimization==0.5.0; python_version >= '3'
tensorflow==2.4.1
termcolor==1.1.0
text-unidecode==1.3
tf-models-official==2.3.0
tf-slim==1.1.0
threadpoolctl==2.1.0; python_version >= '3.5'
tqdm==4.57.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
typeguard==2.11.1; python_full_version >= '3.5.3'
typing-extensions==3.7.4.3; python_version < '3.8'
uritemplate==3.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
werkzeug==1.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
wheel==0.36.2; python_version >= '3'
wrapt==1.12.1
zipp==3.4.0; python_version < '3.8'
